removing old_files from directory
